chore(money-field): move to packages
