+++ packages/components/fields/money-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './money-field';
+
+export { default as version } from './version';
