--- src/components/fields/money-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './money-field';
